requirements.txt
APScheduler==3.7.0
Django==3.0.1
django_recaptcha==3.0.0
matlab==0.1
matplotlib==2.1.2
nibabel==3.2.2
numpy==1.17.4
pandas==1.1.5
paramiko==2.10.3
PyMySQL==0.7.9
pytz==2019.3
scikit_learn==1.0.2
scipy==1.3.3
seaborn==0.8.1
statsmodels==0.12.1
